requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-utils-six==2.0
django-widget-tweaks==1.4.11
idna==3.3
kavenegar==1.1.2
Pillow==9.0.0
python-decouple==3.5
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
